Rename new components  to stable name (#3999)
