+++ packages/client/src/components/reports/vat-monthly-report/misc-table.tsx
@@ -4,7 +4,7 @@ import type { OnChangeFn, RowSelectionState } from '@tanstack/react-table';
 import { Card } from '@/components/ui/card.js';
 import { VatReportMiscTableFieldsFragmentDoc } from '../../../gql/graphql.js';
 import { getFragmentData, type FragmentType } from '../../../gql/index.js';
-import { NewChargesTable } from '../../charges/new-charges-table.js';
+import { ChargesTable } from '../../charges/charges-table.js';
 import { Button } from '../../ui/button.js';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-expressions -- used by codegen
@@ -44,7 +44,7 @@ export const MiscTable = ({ data, rowSelection, onRowSelectionChange }: Props): 
         </h2>
       </div>
       {isOpened && chargesData && (
-        <NewChargesTable
+        <ChargesTable
           data={chargesData.differentMonthDoc}
           rowSelection={rowSelection}
           onRowSelectionChange={onRowSelectionChange}
